converted project files to typescript
