+++ .gitignore
@@ -1,4 +1,6 @@
-node_modules
+node_modules/
+.node_modules/
 config.json
 yarn.lock
 package-lock.json
+built
