--- commands/misc/ping.js
@@ -1,10 +1,0 @@
-const { SlashCommandBuilder } = require('@discordjs/builders');
-
-module.exports = {
-    data: new SlashCommandBuilder()
-        .setName('ping')
-        .setDescription('Replies with Pong!'),
-    async execute(interaction) {
-        await interaction.reply('Pong!');
-    },
-};
