--- events/discord/ready.js
@@ -1,9 +1,0 @@
-const { Events } = require('discord.js');
-
-module.exports = {
-	name: Events.ClientReady,
-	once: true,
-	execute(client) {
-		console.log(`Ready! Logged in as ${client.user.tag}`);
-	},
-};
