+++ package.json
@@ -25,6 +25,7 @@
     "eslint": "^8.55.0",
     "eslint-config-google": "^0.14.0",
     "husky": "^8.0.3",
-    "lint-staged": "^15.2.0"
+    "lint-staged": "^15.2.0",
+    "typescript": "^5.3.3"
   }
 }
