+++ src/deploy-commands.ts
@@ -1,0 +1,45 @@
+const { REST, Routes } = require('discord.js');
+const { clientId, token } = require('../config.json');
+const fs = require('node:fs');
+const path = require('node:path');
+
+const commands = [];
+const basePath = path.join(__dirname, '../');
+const foldersPath = path.join(basePath, 'commands');
+const commandFolders = fs.readdirSync(foldersPath);
+
+for (const folder of commandFolders) {
+    const commandsPath = path.join(foldersPath, folder);
+    const commandFiles = fs.readdirSync(commandsPath).filter(file => file.endsWith('.js'));
+    // Grab the SlashCommandBuilder#toJSON() output of each command's data for deployment
+    for (const file of commandFiles) {
+        const filePath = path.join(commandsPath, file);
+        const command = require(filePath);
+        if ('data' in command && 'execute' in command) {
+            commands.push(command.data.toJSON());
+        }
+        else {
+            console.log(`[WARNING] The command at ${filePath} is missing a required "data" or "execute" property.`);
+        }
+    }
+}
+
+// Construct and prepare an instance of the REST module
+const rest = new REST().setToken(token);
+
+(async () => {
+    try {
+        console.log(`Started refreshing ${commands.length} application (/) commands.`);
+
+        // I guess if we want to do testing, we should add a guild id here to only deploy there
+        const data = await rest.put(
+            Routes.applicationCommands(clientId),
+            { body: commands },
+        );
+
+        console.log(`Successfully reloaded ${data.length} application (/) commands.`);
+    }
+    catch (error) {
+        console.error(error);
+    }
+})();
